Function verification with CBMC

+++ verification/AREA_OF_THE_CIRCLE_THAT_HAS_A_SQUARE_AND_A_CIRCLE_INSCRIBED_IN_IT.c
@@ -9,6 +9,6 @@
 #include <math.h>
 
 float f_gold_c ( int a ) {
-  float area = (M_PI * (float )a * (float)a )/4.0;
+  float area = ( M_PI * a * a ) / 4.0;
   return area;
 }
